backend/requirements.txt
django==3.2.18
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
qrcode[pil]
Pillow
djoser
psycopg2-binary
